[change] Command 'scholar cleanLocal' on all users
The 'scholar cleanLocal' command considers the username argument
optional. If no username argument is provided clean local will
remove extra data for all users on the device. In the GUI doing
clean local on the login page will act as a call with no username
argument, and doing cleann local on any page after the login
will target the logged in user only.